new file:   src/management-system-v2/app/(dashboard)/[environmentId]/machine-config/[configId]/layout.tsx new file:   src/management-system-v2/app/(dashboard)/[environmentId]/machine-config/[configId]/page.tsx new file:   src/management-system-v2/app/(dashboard)/[environmentId]/machine-config/folder/[folderId]/page.tsx new file:   src/management-system-v2/app/(dashboard)/[environmentId]/machine-config/folder/page.tsx modified:   src/management-system-v2/app/(dashboard)/[environmentId]/machine-config/machine-config-list.tsx modified:   src/management-system-v2/app/(dashboard)/[environmentId]/machine-config/page.tsx new file:   src/management-system-v2/components/machine-config-creation-button.tsx modified:   src/management-system-v2/lib/data/legacy/machine-config.ts modified:   src/management-system-v2/lib/data/machine-config-schema.ts
Just starting to figure out the project...
I added the double click functionality to route to a new configuration page and added the basic layout for the configuration page.
I also added the folder page and the layout for the folder page.
I also added the machine-config-creation-button component to create a new machine configuration.
All of these are unfinished and need more work.